create policy for permission, patient history and patient show with all data

+++ app/Http/Controllers/Backend/ConsultantController.php
@@ -12,56 +12,34 @@ use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Storage;
 
 class ConsultantController extends Controller
-{ 
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    function __construct()
-    {
-         $this->middleware('permission:consultant-list|consultant-create|consultant-edit|consultant-delete', ['only' => ['index','show']]);
-         $this->middleware('permission:consultant-create', ['only' => ['create','store']]);
-         $this->middleware('permission:consultant-edit', ['only' => ['edit','update']]);
-         $this->middleware('permission:consultant-delete', ['only' => ['destroy']]);
-    }
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+{
     public function index(Request $request): View|string
     {
-        $pageTitle = "Consultants List";
+        $this->authorize('viewAny', Consultant::class);
 
+        $pageTitle = "Consultants List";
         $consultants = Consultant::latest()->paginate(5);
+
         if ($request->ajax()) {
             return view('consultants.list', compact('consultants'))->render();
         }
 
-        return view('consultants.index',compact('consultants','pageTitle'));
+        return view('consultants.index', compact('consultants', 'pageTitle'));
     }
-    
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function create(): View
     {
+        $this->authorize('create', Consultant::class);
+
         $pageTitle = "Consultants Create";
         $insurances = Insurance::all();
-        return view('consultants.create', compact('pageTitle','insurances'));
+        return view('consultants.create', compact('pageTitle', 'insurances'));
     }
-    
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
+
     public function store(ConsultantRequest $request): JsonResponse
     {
+        $this->authorize('create', Consultant::class);
+
         $data = $request->validated();
 
         if ($request->hasFile('image')) {
@@ -69,7 +47,6 @@ class ConsultantController extends Controller
         }
 
         $consultant = Consultant::create($data);
-
         $consultant->insurances()->sync($request->input('insurance_id', []));
 
         return response()->json([
@@ -77,42 +54,28 @@ class ConsultantController extends Controller
             'message' => 'Consultant created successfully',
         ]);
     }
-    
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\consultant  $consultant
-     * @return \Illuminate\Http\Response
-     */
+
     public function show(Consultant $consultant): View
     {
+        $this->authorize('view', $consultant);
+
         $pageTitle = "Show Consultant";
-        return view('consultants.show',compact('consultant','pageTitle'));
+        return view('consultants.show', compact('consultant', 'pageTitle'));
     }
-    
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\consultant  $consultant
-     * @return \Illuminate\Http\Response
-     */
+
     public function edit(Consultant $consultant): View
     {
+        $this->authorize('update', $consultant);
+
         $pageTitle = "Edit Consultant";
         $insurances = Insurance::all();
-
-        return view('consultants.edit',compact('consultant','pageTitle','insurances'));
+        return view('consultants.edit', compact('consultant', 'pageTitle', 'insurances'));
     }
-    
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\consultant  $consultant
-     * @return \Illuminate\Http\Response
-     */
+
     public function update(ConsultantRequest $request, Consultant $consultant): JsonResponse
     {
+        $this->authorize('update', $consultant);
+
         $data = $request->validated();
 
         if ($request->hasFile('image')) {
@@ -125,8 +88,6 @@ class ConsultantController extends Controller
         }
 
         $consultant->update($data);
-
-        // Sync selected insurance IDs
         $consultant->insurances()->sync($request->input('insurance_id', []));
 
         return response()->json([
@@ -135,17 +96,12 @@ class ConsultantController extends Controller
         ]);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Consultant  $consultant
-     * @return \Illuminate\Http\Response
-     */
     public function destroy(Consultant $consultant): RedirectResponse
     {
+        $this->authorize('delete', $consultant);
+
         $consultant->delete();
-    
-        return redirect()->route('consultants.index')
-                        ->with('success','Consultant deleted successfully');
+
+        return redirect()->route('consultants.index')->with('success', 'Consultant deleted successfully');
     }
 }
